split out cors, entry fetching concerns
