+++ src/frontend/concerns.rs
@@ -1,15 +1,28 @@
 //! Common concerns for frontend routes
-use crate::backend::Backend;
-use crate::backend::RerunOptions;
-use crate::frontend::cached::task_report;
-use crate::frontend::helpers::*;
-use crate::frontend::params::{ReportParams, RerunRequestParams, TemplateContext};
-use crate::models::{Corpus, HistoricalRun, Service};
+use redis::Commands;
+use regex::Regex;
 use rocket::request::Form;
 use rocket::response::status::{Accepted, NotFound};
+use rocket::response::{NamedFile, Redirect};
+use rocket::Data;
 use rocket_contrib::json::Json;
 use rocket_contrib::templates::Template;
 use std::collections::HashMap;
+use std::str;
+
+use crate::backend::Backend;
+use crate::backend::RerunOptions;
+use crate::frontend::cached::task_report;
+use crate::frontend::captcha::{check_captcha, safe_data_to_string};
+use crate::frontend::helpers::*;
+use crate::frontend::params::{ReportParams, RerunRequestParams, TemplateContext};
+use crate::models::{Corpus, HistoricalRun, Service, Task};
+
+lazy_static! {
+  static ref STRIP_NAME_REGEX: Regex = Regex::new(r"/[^/]+$").unwrap();
+}
+/// Placeholder word for unknown filters/fields
+pub const UNKNOWN: &str = "_unknown_";
 
 /// Prepare a configurable report for a <corpus,server> pair
 pub fn serve_report(
@@ -282,4 +295,175 @@ pub fn serve_rerun(
     Err(_) => Err(NotFound("Access Denied".to_string())), // TODO: better error message?
     Ok(_) => Ok(Accepted(None)),
   }
+}
+
+/// Provide a `NamedFile` for an entry, redirecting if captcha guard isn't met
+pub fn serve_entry(
+  service_name: String,
+  entry_id: usize,
+  data: Data,
+) -> Result<NamedFile, Redirect>
+{
+  // Any secrets reside in config.json
+  let cortex_config = load_config();
+  let data = safe_data_to_string(data).unwrap_or_default(); // reuse old code by setting data to the String
+  let g_recaptcha_response_string = if data.len() > 21 {
+    let data = &data[21..];
+    data.replace("&g-recaptcha-response=", "")
+  } else {
+    UNKNOWN.to_owned()
+  };
+  let g_recaptcha_response = &g_recaptcha_response_string;
+  // Check if we hve the g_recaptcha_response in Redis, then reuse
+  let mut redis_opt;
+  let quota: usize = match redis::Client::open("redis://127.0.0.1/") {
+    Err(_) => return Err(Redirect::to("/")), // TODO: Err(NotFound(format!("redis unreachable")))},
+    Ok(redis_client) => match redis_client.get_connection() {
+      Err(_) => return Err(Redirect::to("/")), /* TODO: Err(NotFound(format!("redis
+                                                 * unreachable")))}, */
+      Ok(mut redis_connection) => {
+        let quota = redis_connection.get(g_recaptcha_response).unwrap_or(0);
+        redis_opt = Some(redis_connection);
+        quota
+      },
+    },
+  };
+
+  println!("Response: {:?}", g_recaptcha_response);
+  println!("Quota: {:?}", quota);
+  let captcha_verified = if quota > 0 {
+    if let Some(ref mut redis_connection) = redis_opt {
+      println!("Using local redis quota.");
+      if quota == 1 {
+        // Remove if last
+        redis_connection.del(g_recaptcha_response).unwrap_or(());
+      } else {
+        // We have quota available, decrement it
+        redis_connection
+          .set(g_recaptcha_response, quota - 1)
+          .unwrap_or(());
+      }
+      // And allow operation
+      true
+    } else {
+      false // no redis, no access.
+    }
+  } else {
+    // expired quota, check with google
+    let check_val = check_captcha(g_recaptcha_response, &cortex_config.captcha_secret);
+    println!("Google validity: {:?}", check_val);
+    if check_val {
+      if let Some(ref mut redis_connection) = redis_opt {
+        // Add a reuse quota if things check out, 19 more downloads
+        redis_connection.set(g_recaptcha_response, 19).unwrap_or(());
+      }
+    }
+    check_val
+  };
+  println!("Captcha validity: {:?}", captcha_verified);
+
+  // If you are not human, you have no business here.
+  if !captcha_verified {
+    if g_recaptcha_response != UNKNOWN {
+      return Err(Redirect::to("/expire_captcha"));
+    } else {
+      return Err(Redirect::to("/"));
+    }
+  }
+
+  let backend = Backend::default();
+  match Task::find(entry_id as i64, &backend.connection) {
+    Ok(task) => {
+      let entry = task.entry;
+      let zip_path = match service_name.as_str() {
+        "import" => entry,
+        _ => STRIP_NAME_REGEX.replace(&entry, "").to_string() + "/" + &service_name + ".zip",
+      };
+      if zip_path.is_empty() {
+        Err(Redirect::to("/")) // TODO : Err(NotFound(format!("Service {:?} does not have a result
+                               // for entry {:?}", service_name,
+                               // entry_id)))
+      } else {
+        NamedFile::open(&zip_path).map_err(|_| Redirect::to("/"))
+      }
+    },
+    Err(e) => {
+      dbg!(e); // TODO: Handle these better
+      Err(Redirect::to("/"))
+    },
+  }
+}
+
+/// Serves an entry as a `Template` instance to be preview via a client-side asset renderer
+pub fn serve_entry_preview(
+  corpus_name: String,
+  service_name: String,
+  entry_name: String,
+) -> Result<Template, NotFound<String>>
+{
+  let report_start = time::get_time();
+  let corpus_name = corpus_name.to_lowercase();
+  let mut context = TemplateContext::default();
+  let mut global = HashMap::new();
+  let backend = Backend::default();
+
+  let corpus_result = Corpus::find_by_name(&corpus_name, &backend.connection);
+  if let Ok(corpus) = corpus_result {
+    let service_result = Service::find_by_name(&service_name, &backend.connection);
+    if let Ok(service) = service_result {
+      // Assemble the Download URL from where we will gather the page contents (after captcha is
+      // confirmed) First, we need the taskid
+      let task = match Task::find_by_name(&entry_name, &corpus, &service, &backend.connection) {
+        Ok(t) => t,
+        Err(e) => return Err(NotFound(e.to_string())),
+      };
+      let download_url = format!("/entry/{}/{}", service_name, task.id.to_string());
+      global.insert("download_url".to_string(), download_url);
+
+      // Metadata for preview page
+      global.insert(
+        "title".to_string(),
+        "Corpus Report for ".to_string() + &corpus_name,
+      );
+      global.insert(
+        "description".to_string(),
+        "An analysis framework for corpora of TeX/LaTeX documents - statistical reports for "
+          .to_string()
+          + &corpus_name,
+      );
+      global.insert("corpus_name".to_string(), corpus_name);
+      global.insert("corpus_description".to_string(), corpus.description);
+      global.insert("service_name".to_string(), service_name);
+      global.insert(
+        "service_description".to_string(),
+        service.description.clone(),
+      );
+      global.insert("type".to_string(), "Conversion".to_string());
+      global.insert("inputformat".to_string(), service.inputformat.clone());
+      global.insert("outputformat".to_string(), service.outputformat.clone());
+      match service.inputconverter {
+        Some(ref ic_service_name) => {
+          global.insert("inputconverter".to_string(), ic_service_name.clone())
+        },
+        None => global.insert("inputconverter".to_string(), "missing?".to_string()),
+      };
+      global.insert("report_time".to_string(), time::now().rfc822().to_string());
+    }
+    global.insert("severity".to_string(), entry_name.clone());
+    global.insert("entry_name".to_string(), entry_name);
+  }
+
+  // Pass the globals(reports+metadata) onto the stash
+  context.global = global;
+  // And pass the handy lambdas
+  // And render the correct template
+  decorate_uri_encodings(&mut context);
+
+  // Report also the query times
+  let report_end = time::get_time();
+  let report_duration = (report_end - report_start).num_milliseconds();
+  context
+    .global
+    .insert("report_duration".to_string(), report_duration.to_string());
+  Ok(Template::render("task-preview", context))
 }
