+++ src/frontend/cors.rs
@@ -1,0 +1,38 @@
+//! CORS capabilities for the Rocket frontend
+use rocket::fairing::{Fairing, Info, Kind};
+use rocket::http::Header;
+use std::io::Cursor;
+/// Rocket solution for Cross-origin resource sharing
+pub struct CORS();
+
+impl Fairing for CORS {
+  fn info(&self) -> Info {
+    Info {
+      name: "Add CORS headers to requests",
+      kind: Kind::Response,
+    }
+  }
+
+  fn on_response(&self, request: &rocket::Request, response: &mut rocket::Response) {
+    if request.method() == rocket::http::Method::Options
+      || response.content_type() == Some(rocket::http::ContentType::JSON)
+    {
+      response.set_header(Header::new("Access-Control-Allow-Origin", "*"));
+      response.set_header(Header::new(
+        "Access-Control-Allow-Methods",
+        "POST, GET, OPTIONS",
+      ));
+      response.set_header(Header::new("Access-Control-Allow-Headers", "Content-Type"));
+      response.set_header(Header::new("Access-Control-Allow-Credentials", "true"));
+      response.set_header(Header::new(
+        "Content-Security-Policy-Report-Only",
+        "default-src https:; report-uri /csp-violation-report-endpoint/",
+      ));
+    }
+
+    if request.method() == rocket::http::Method::Options {
+      response.set_header(rocket::http::ContentType::Plain);
+      response.set_sized_body(Cursor::new(""));
+    }
+  }
+}
