+++ src/frontend/mod.rs
@@ -2,5 +2,6 @@
 pub mod cached;
 pub mod captcha;
 pub mod concerns;
+pub mod cors;
 pub mod helpers;
 pub mod params;
